pass DBOptions and log_number to log::Writer and log::Reader
The writer needs to know what the db options are in order to know whether to
use the new or old CRC scheme.  We also need to know the log number in the
new case.  Likewise, the reader will need the same info to validate the
CRC.

Note that in the reader case the DBOptions are optional, as not all callers
have it available.  It's helpful for debugging in the general case, though.

Signed-off-by: Sage Weil <sage@redhat.com>