db/log_reader: treat bad record length or checksum as kEof
If we are in kTolerateCorruptedTailRecords, treat these errors as the
end of the log.  This is particularly important for recycled logs, where
we will regularly see corrupted headers (bad length or checksum) when
replaying a log.  If we are aligned with a block boundary or get lucky,
we will land on an old header and see the log nubmer mismatch, but more
commonly we will land midway through some previously and effectively
see noise.

Signed-off-by: Sage Weil <sage@redhat.com>